refactor(client): move type definitions out of utils/composables into app/types/
CLAUDE.md Rule #1 ("NEVER Define Types/Interfaces Inline in Vue Files")
applies to `composables/**/*.ts` and `utils/**/*.ts` by the same logic
that motivates the vue-file ban — types should be discoverable in
`app/types/` regardless of which consumer they support.

caught two violators introduced during the admin-config + friendly-error
work:

- `utils/api/admin-mcp/friendly-error.ts` declared `OfetchLikeError`
  inline. moved to new `app/types/fetch.ts` so any future `$fetch`
  consumer can import the same narrowing shape without recreating it.
- `composables/admin/config-schema.ts` declared `ConfigFieldType`,
  `ConfigFieldSchema`, `ConfigSectionSchema`; `composables/admin/use-
  admin-config.ts` declared `ConfigLine`, `ConfigSectionView`. moved
  all five to new `app/types/admin-config-schema.ts`, distinct from the
  existing `admin-config.ts` (which mirrors the live API response,
  while the schema catalog describes the keys you *could* set).

`composables/admin/config-schema.ts` now only holds the catalog data
(`CONFIG_SCHEMA: readonly ConfigSectionSchema[]`); the type definitions
live in `~/types/admin-config-schema`. `use-admin-config.ts` imports
the four type names it actually uses (`AdminBreadcrumbCrumb`,
`ConfigLine`, `ConfigSectionSchema`, `ConfigSectionView`) from
`~/types` and the catalog from `./config-schema`.

types/index.ts barrel updated to re-export the new modules.

post-refactor lint + nuxt build green.
